feat(stimuli)!: draw the SD card's localization from the fitted rhythm too
Closes the seam the previous commit left open. The fitted resting rhythm is now the
default and ONLY way this repo generates a localization pulse: the RC device runs it
live, and the SD card gets it through the library's 6 synthetic localization items,
which `generate_localization` draws from the model instead of from the retired
lognormal-around-a-target-rate ladder.

Re-exported against the mounted FLONA 2025 dataset. Invariant 3 holds exactly: EOD_HV,
the 5 real volleys, both real localization exemplars and all 100 synthetic volleys came
back byte-identical, and precisely the 6 synthetic localization items (107-112) moved.
The decoupled RNG streams did their job — a localization-side change moved no volley.

LIBRARY FORMAT v3 -> v4: StimItem.ipi_samp widened uint16 -> uint32. The old ceiling was
65535 samples == 1.31 s, and that was not a tail limit but a core-of-the-distribution
one: at the 1 Hz rung 35 % of the model's intervals are longer than that, and even at
10 Hz 1.9 % are. Clamping would have deformed every item at every rung, so the format
holds the model rather than the model being bent to fit the format. +2 bytes per pulse;
the packet is 58.7 kB of a 131 kB budget. `packet_bytes` now reads the widths off the
arrays' own dtypes, because a hard-coded `* 2` there under-reported by 22 kB against
`max_flash_bytes` — the one number that is supposed to catch a library outgrowing flash.

Three things bit during the port and are now documented rather than merely fixed:

* A FINITE ITEM CANNOT HOLD A LABELLED TEMPO. A 60 s item — of which program B renders
  only the first 20 s — is far too short to average out components relaxing over 96 s
  and 62 min. Left alone a "5 Hz" rung realises anywhere in 1.8-10.3 Hz and the rungs
  come out in the wrong ORDER. Each train is therefore time-scaled onto its rung, which
  is legitimate because rate is a pure time dilation: CV2 and the autocorrelation are
  scale-invariant, so only the label changes. Grow THEN scale — scaling then trimming
  hands the item a prefix whose median is something else again, which shipped a "7 Hz"
  rung ticking at 11 Hz before it was caught.
* `_longest_gapfree` MUST NOT TOUCH A SYNTHETIC TRAIN. It is a data-quality gate for
  real exemplars, where a multi-second hole is ambiguous between a quiet fish and a lost
  track. Applied to model output it deletes exactly the long silences the model was
  adopted for; it chopped a 37-pulse train to 14.
* THE ANTIMODE IS ABSOLUTE BIOLOGY. The model's interval table is clamped to the 25 ms
  resting/volley antimode in SCORE space, but rate is a TIME dilation, so a fast tempo
  scales that floor down with everything else — 8.8 ms at the 10 Hz rung, with 7 % of
  "localization" intervals landing inside volley territory. The source analysis segments
  resting from fast runs on exactly this antimode, so those pairs would read as
  micro-bursts of our own making. Both paths now re-impose it in absolute time, and
  LOC_REFRACTORY_SAMP is the antimode (1250 samp = 25 ms) rather than the arbitrary 5 ms
  it was — which sat squarely inside volley territory. This is a LOWER clamp and is not
  in tension with leaving the long tail alone; the spec clamps both ends of the table and
  defends the bottom one in these terms.

A rung is now a TICK TEMPO, not an average pulse rate, matching CH3 on the remote: the
3 Hz rung delivers ~1.9 pulses/s. Program D's 5 s lead gets only ~2 pulses from the 1 Hz
rung — a 1 Hz fish with the model's tail, not a defect.

The eel_player self-test's runaway guard was a flat 60 s annotated "no case here is
anywhere near that", which stopped being true when the localization items grew past 80 s.
It is derived from the item's own cumulative IPI now, so it cannot go stale again.

make check is green, both Teensy targets included. The WAV card needs rebuilding with
fakefish-build-card before the next outing — see TODO.md 6b.

+++ firmware/eel_fakefish_rc/eel_fakefish_rc.ino
@@ -142,7 +142,9 @@ static uint32_t   g_sham_phase = 0;                  // drives the distinct SHAM
 static_assert(PULSE_MARKER_IPI_SAMP >= EEL_PLAYER_MIN_IPI_SAMP,
               "PULSE_MARKER_IPI_SAMP is tighter than eel_player can mix — raise the marker's "
               "IPI, or widen EEL_PLAYER_MIN_IPI_SAMP and pay for it on every ISR tick");
-static uint16_t   g_marker_ipi[PULSE_MARKER_MAX_PULSES];
+// uint32_t, matching StimItem::ipi_samp since library format v4 widened it (the
+// localization items now carry multi-second silences that a uint16 cannot hold).
+static uint32_t   g_marker_ipi[PULSE_MARKER_MAX_PULSES];
 static StimItem   g_marker_item;
 
 // ===== pulse log (ISR-owned counters; the FILE is owned by loop()) ========
